adapt functions

+++ anRpackage/src/main.c
@@ -131,7 +131,6 @@ void setParameters(Matrix *x, Matrix *w)
  * - `x` and `w` dimensions must be coherent.
  *
  */
-/*
 Matrix getInitialP(const char *p_method)
 {
 
@@ -160,9 +159,8 @@ Matrix getInitialP(const char *p_method)
     else if (strcmp(p_method, "proportional") == 0)
     {
         for (int c = 0; c < TOTAL_CANDIDATES; c++)
-        { // --- For each candidate
-            double ratio =
-                (double)CANDIDATES_VOTES[c] / (double)TOTAL_VOTES; // Proportion of candidates votes per total votes.
+        {                                                     // --- For each candidate
+            double ratio = CANDIDATES_VOTES[c] / TOTAL_VOTES; // Proportion of candidates votes per total votes.
             for (int g = 0; g < TOTAL_GROUPS; g++)
             { // --- For each group, given a candidate
                 MATRIX_AT(probabilities, g, c) = ratio;
@@ -223,7 +221,6 @@ Matrix getInitialP(const char *p_method)
     // ---...--- //
     return probabilities;
 }
-*/
 /*
  * @brief Computes the log-likelihood for a given probability and `q` array
  *
